Q35: Don't Compress EFI_APPLICATIONs
During the initial development of Patina, EFI_APPLICATIONs
were moved to be compressed. This is an unusual configuration
and is not required. In addition, Q35 was not dispatching
the compressed EFI_APPLICATIONs (being followed up with separately).

This commit removes compressing EFI_APPLICATIONs and now dispatches
them correctly.

This also increases the DXE FV size, because the GCC build expands
past the FV boundary with this change.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>